feat: add input validation for graphql backend (#1437)
Input validation can be done at: (from
https://stackoverflow.com/a/59682946)
- in resolver with the general approach, e.g. if (validationPass) {...}
else {..} - Too verbose, keep the resolver thin is a best practice.
Because It's some kind of route layer. Not recommend
- in resolver using a composition library - Recommend
- in the model layer (ORM) using some decorators or model schema
definition. - This is the traditional effective way no matter you use
GraphQL, gRPC, soap WSDL, RESTful API. From the classic MVC software
development idea. Recommend
- in GraphQL schema directive - A little complex, not recommend
- in GraphQL
[middleware](https://www.npmjs.com/package/graphql-middleware) - A
little complex, you need to design the middleware and map the validation
rules to the corresponding Query and Mutation, you quite possibly need
to use the info parameter to get the GraphQL operation. Not recommend

Prisma currently doesn't provide a built-in way, it amounts to calling
the input verification explicitly, see eg
https://www.prisma.io/docs/concepts/components/prisma-client/custom-validation.
So validation at the ORM layer is not very convenient.

The approach we take here is a middleground of the first two approaches,
and we declare the validation as a typescript decorator. Moreover, the
validation schema are defined in a central place, so that they can be
easily reused on the frontend. This is preparation for #139.